move concrete to utils
